[Bug/#50] 회차별 정보 추출 시 리사이클러뷰가 조회되지 않는 문제

+++ app/src/main/java/com/example/momolabfe/data/remote/record/model/RecordResponse.kt
@@ -25,7 +25,7 @@ data class RecordExchangeOcrResponse (
     @SerializedName("exchange_time") val exchangeTime: LocalTime,
     @SerializedName("drain_volume") val drainVolume: Int,
     @SerializedName("fill_volume") val fillVolume: Int,
-    @SerializedName("fill_concentration") val fillConcentration: Float,
+    @SerializedName("fill_concentration") val fillConcentration: Double,
     @SerializedName("uf") val uf: Int
 )
